+++ edison-core/src/main/java/de/otto/edison/metrics/configuration/MetricsJvmConfiguration.java
@@ -1,12 +1,12 @@
 package de.otto.edison.metrics.configuration;
 
 import com.codahale.metrics.MetricRegistry;
-import com.codahale.metrics.jvm.FileDescriptorRatioGauge;
-import com.codahale.metrics.jvm.GarbageCollectorMetricSet;
-import com.codahale.metrics.jvm.MemoryUsageGaugeSet;
+import com.codahale.metrics.jvm.*;
 import com.ryantenney.metrics.spring.config.annotation.EnableMetrics;
 import com.ryantenney.metrics.spring.config.annotation.MetricsConfigurerAdapter;
 import org.springframework.context.annotation.Configuration;
+
+import java.util.concurrent.TimeUnit;
 
 @Configuration
 @EnableMetrics
@@ -17,6 +17,7 @@ public class MetricsJvmConfiguration extends MetricsConfigurerAdapter {
         metricRegistry.register("gc", new GarbageCollectorMetricSet());
         metricRegistry.register("memory", new MemoryUsageGaugeSet());
         metricRegistry.register("filedescriptors.ratio", new FileDescriptorRatioGauge());
+        metricRegistry.register("threads", new CachedThreadStatesGaugeSet(10, TimeUnit.SECONDS));
     }
 
 }